take responsive image functionality and turn it into a ResponsiveImage component
