+++ router/nginx.conf
@@ -90,7 +90,7 @@ http {
       proxy_set_header        Host $host;
       proxy_pass              http://v2;
     }
-    location ~ ^/exhibitions/W|^/exhibitions {
+    location ~ ^/exhibitions/W|^/exhibitions$ {
       proxy_set_header        Host $host;
       proxy_pass              http://v2;
     }
